Update the TS typings
